// src/app/models/Ouststanding/out-standing-detail.model.ts
export class OutStandingDetail {
    companyId:    number;
    companyName:  string;
    partyName:    string;
    partyId:      number;
    agentId:      number;
    agentName:    string;
    BillNo:       number;
    BillDate:     Date;
    BillAmount:   number;
    PaidAmount:   number;
    Balance:      number;
    Id:           number;
    date1:     Date;
    date2:     Date;

}